fix pypi tests
Former-commit-id: 47d6dc0b777bb3d8817497edaf52f8f3cfda601c

+++ river/metrics/geometric_mean.py
@@ -51,7 +51,7 @@ class GeometricMean(base.MultiClassMetric):
     def get(self):
 
         if self.cm.n_classes > 0:
-            sensitivity_per_class = np.zeros(self.cm.n_classes, np.float)
+            sensitivity_per_class = np.zeros(self.cm.n_classes, float)
             for i, c in enumerate(self.cm.classes):
                 try:
                     sensitivity_per_class[i] = self.cm[c][c] / self.cm.sum_row[c]
